be/services/matchMakingServices/requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
redis
requests
pyjwt
python-dotenv
gevent
flask-cors
